ai-elderly-medicare-system/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-elderly-medicare-system"
version = "1.0.0"
description = "A comprehensive healthcare management system for elderly patients with AI-powered features"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "AI Medicare Team", email = "support@ai-medicare-system.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.7.0",
    "pydantic>=1.8.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.5",
    "email-validator>=1.1.3",
    "celery>=5.1.0",
    "redis>=3.5.3",
    "requests>=2.25.1",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "tensorflow>=2.6.0",
    "openai>=0.27.0",
    "twilio>=7.0.0",
    "python-dotenv>=0.19.0",
    "pydantic[dotenv]>=1.8.0",
    "streamlit>=1.0.0",
    "plotly>=5.0.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "jupyter>=1.0.0",
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
]

[project.optional-dependencies]
dev = [
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/your-organization/ai-elderly-medicare-system"
Documentation = "https://github.com/your-organization/ai-elderly-medicare-system/docs"
Repository = "https://github.com/your-organization/ai-elderly-medicare-system.git"
"Bug Tracker" = "https://github.com/your-organization/ai-elderly-medicare-system/issues"

[project.scripts]
medicare-run = "backend.app.main:main"
medicare-init = "scripts.setup_database:main"
medicare-seed = "scripts.seed_data:main"

[tool.setuptools.packages.find]
where = ["backend"]
include = ["app*", "app.*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*", "*Test"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "pydantic.mypy"
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true